% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstan_evidence_calculator.R
\name{Evidencerun}
\alias{Evidencerun}
\title{calculates the Evidence from samplings at different beta values}
\usage{
Evidencerun(rstanrunvector, betavector, logLiklabel = "loglikelihood")
}
\arguments{
\item{rstanrunvector}{a vector of rstan samplings obtained from runSampling}

\item{betavector}{the vector of beta values corresponding to rstanrunvector}

\item{logLiklabel}{the label in the model that is used for loglikelihood values}
}
\description{
calculates the Evidence from samplings at different beta values
}
