% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstan_evidence_calculator.R
\name{runSampling}
\alias{runSampling}
\title{runs a the stan sampling of a model for some data and a vector of beta values}
\usage{
runSampling(model, mydata, betavector, betalabel = "beta", ...)
}
\arguments{
\item{model}{a rstan model}

\item{mydata}{a named list providing the data for the stan model}

\item{betavector}{a vector of values between 0 and 1 for beta}

\item{betalabel}{the label in the model that is used fot beta}
}
\description{
runs a the stan sampling of a model for some data and a vector of beta values
}
